let autoboxing do the work
git-svn-id: https://svn.apache.org/repos/asf/click/trunk@942900 13f79535-47bb-0310-9956-ffa450edef68

+++ click/framework/src/org/apache/click/ClickServlet.java
@@ -1374,7 +1374,7 @@ public class ClickServlet extends HttpServlet {
                 response.setDateHeader(name, time);
 
             } else {
-                int intValue = ((Integer) value).intValue();
+                int intValue = (Integer) value;
                 response.setIntHeader(name, intValue);
             }
         }
